backend/api/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
pydantic[email]>=2.4.0
python-multipart>=0.0.6
httpx>=0.25.0
loguru>=0.7.2
jinja2>=3.1.2

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.12.0
geoalchemy2>=0.14.0
geohash2>=1.1
geopy>=2.4.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
# Pin bcrypt to a version compatible with passlib 1.7.x to avoid AttributeError: module 'bcrypt' has no attribute '__about__'
bcrypt==4.0.1

# Task Queue
celery>=5.3.4
redis>=5.0.0

# Cloud Storage
boto3>=1.28.0

# Notifications
sendgrid>=6.11.0
twilio>=8.10.0
firebase-admin>=6.2.0

# Monitoring
sentry-sdk[fastapi]>=1.38.0
prometheus-client>=0.19.0

# Rate Limiting
slowapi>=0.1.9

# Configuration
python-decouple>=3.8

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
requests>=2.31.0
